changed folder structure and added composer
