+++ composer.json
@@ -1,0 +1,9 @@
+{
+    "name": "delphiki/subrip-file-parser",
+    "license": "BSD",
+    "type": "project",
+    "description": "A library for SubRip (srt) files manipulation",
+    "autoload": {
+        "psr-0": { "SrtParser": "src/" }
+    }
+}
